Very basic Markdown parsing for recipe/ingredients
Supports overall file structure (recipe name, ingredients, steps) with
Markdown headings and simple list parsing for ingredients. An ingredient
has a name, a quantity, an optional unit, and an optional list of
attributes.

This project is mostly meant as a library, therefore `main.rs` is also
deleted in favor of `lib.rs`.