Better unit system with various unit types
Units of a certain category (e.g. mass) are now their distinct type
implementing a common trait (`UnitTrait`). The `Unit` enum represents a
unit belonging to a particular category, each of which maps one-to-one
with enum members named like the categories. Additionally, `Unit` can
hold a `Custom` unit which is not recognized as belonging to any of the
categories.

This also introduces the `Quantity` and `QuantityOf<UnitType>` types,
representing a certain amount of, respectively, some arbitrary unit and
some unit of a certain category.

Adds many unit tests for unit/quantity parsing logic.